Support for DateTime64 values before Unix epoch

+++ ClickHouse.Client.Tests/MathUtilsTests.cs
@@ -14,7 +14,7 @@ namespace ClickHouse.Client.Tests
             for (int i = 1; i <= powersMax; i++)
             {
                 value *= (ulong)integer;
-                Assert.AreEqual(value, MathUtils.ToPower((uint)integer, i));
+                Assert.AreEqual(value, MathUtils.ToPower(integer, i));
             }
         }
     }
